requirements.txt
cython==0.25.1
ql-cystatsd==1.1.0
